CLI Options and Zotero Plugin Framework
Options currently allow for setting a debug flag.
The plugin is very rough, and structured for testing purposes at the moment. It is written in typescript using the `zotero-types` package. Uses Webpack to build into a single Javascript file, and then a script to produce ZIP archives for Zotero installation (uses the `xpi` extension).