Replace Plugin UI Injection with Context Menu API
Uses the Zotero reader API to create an export menu item in the reader context menu instead of the main window menus. This is actually a far worse user experience (you must be in the reader to export), but the previous system failed on some OSs (notably Mac) and was much more error prone given the lack of any good general UI APIs to use.