Improved Error Handling in Main
All error sites in main now properly print error messages and return `ProgramError`s, except for those related to global initialization (failure is not fatal in these cases).
To properly deal with very early errors (before even the config is loaded), initialization is split into two stages. The first initializes the panic hook to reasonable defaults, and the second overwrites it with config-provided values. Log files are now always generated next to the executable rather than in the data directory.